[Feat] Create CustomUserDetails
